If you're given a list of hosts, take the car, not cadr
Gnus, at least, provides a single-item list for the host.  To get the
first item one should use `car`; `cadr`, on the other hand, attempts to
get the first item of the tail of the list, a.k.a, the second item.

This is still just a stopgap---according to the documentation of
`auth-source-search`, it should be iterating over that, and perhaps
other entries, not just taking the first.