Host can be in the entry, not only the filename
This patch makes sure an hostname is searched inside entries not only in
entry filenames. This is necessary because some hostnames can't be
encoded in a filename (typically, hostnames with slashes).

Problem of this patch: it makes finding a password slower because each
entry must now be opened. A workaround could be to first search inside
entry filenames for a match.