Search hostname:port/username
According to README.md auth-source-pass supports entries with username
either prefixed to the hostname with an @ as separator or in a
subdirectory under the hostname.  This was true when there was no port
or service included in the name, but not when there was one.

This commit adds a check for the missed case. I believe that it would
work just as well to replace the nil passed for the user in the call
two lines down, but I wasn't completely sure I wasn't missing
something for why nil was passed in that case so chose the
conservative option of adding the more specific check for the case
when both user and port are supplied.