Split out the attribute retrieval form auth-source-pass-get
This commit splits out the retrieval of an attribute value from the
data for an entry in the function `auth-source-pass-get` into a
separate function named `auth-source-pass--get-attr` and uses the new
function in `auth-source-pass--build-result` to eliminate the need to
repeatedly retrieve and parse the data for the entry. This is
generally a good thing since it eliminates repetitions of the same
crypto and parsing operations. It is especially valuable when
protecting an entry with a yubikey with touch required for crypto
operations as it eliminates the need to touch the yubikey sensor for
each attribute retrieved.