Remove ...--select-one-entry and ...--entry-valid-p
These were not being used any longer. I had meant to remove
auth-source-pass--select-one-entry in the original commit. I thought
auth-source-pass--entry-valid-p might be worth keeping around in case
we wanted to test validity in the future, but it is trivial to
re-implement if needed.

The test auth-source-pass-entry-is-not-valid-when-unreadable was also
removed because it was just testing auth-source-pass--entry-valid-p.