Support multiple hosts in search spec
If host in the spec is a list, it tries to find matched entry for the
first element of the list and then tries the next element, and so on.

If "host" field exists in the matched password file, the value is used
instead of that given in the spec.